+++ tests/dryrun/test006--port.sh
@@ -56,6 +56,7 @@ docker \\
     --label 'cb.keep-alive=false' \\
     --label 'cb.daemon=false' \\
     -p 127.0.0.1:${PORT}:10000 \\
+    --add-host host.docker.internal:host-gateway \\
     -e 'BOOTH_SETUPS=/opt/codingbooth/setups' \\
     -e 'BOOTH_CONTAINER_NAME=dryrun' \\
     -e 'BOOTH_DAEMON=false' \\
@@ -85,6 +86,8 @@ docker \\
     -e 'BOOTH_ENV_FILE=' \\
     -e 'BOOTH_HOST_UID=${HOST_UID}' \\
     -e 'BOOTH_HOST_GID=${HOST_GID}' \\
+    -e 'BOOTH_HOST_NAME=host.docker.internal' \\
+    -e 'BOOTH_HOST_IP=XXXXX' \\
     '--pull=never' \\
     -e 'TZ=America/Toronto' \\
     nawaman/codingbooth:base-${VERSION} \\