+++ tests/dryrun/test011-variant.sh
@@ -84,6 +84,7 @@ docker \\
     --label 'cb.keep-alive=false' \\
     --label 'cb.daemon=false' \\
 ${SHM}    -p 127.0.0.1:10000:10000 \\
+    --add-host host.docker.internal:host-gateway \\
     -e 'BOOTH_SETUPS=/opt/codingbooth/setups' \\
     -e 'BOOTH_CONTAINER_NAME=dryrun' \\
     -e 'BOOTH_DAEMON=false' \\
@@ -113,6 +114,8 @@ ${SHM}    -p 127.0.0.1:10000:10000 \\
     -e 'BOOTH_ENV_FILE=' \\
     -e 'BOOTH_HOST_UID=${HOST_UID}' \\
     -e 'BOOTH_HOST_GID=${HOST_GID}' \\
+    -e 'BOOTH_HOST_NAME=host.docker.internal' \\
+    -e 'BOOTH_HOST_IP=XXXXX' \\
     '--pull=never' \\
     -e 'TZ=America/Toronto' \\
     nawaman/codingbooth:${GOT_VARIANT}-${VERSION} \\